requirements-bot.txt
# YouTube Storage Discord Bot dependencies
discord.py>=2.3.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
